Dev to master

+++ src/modules/automovel/useCases/deleteTipoAutomovel/DeleteTipoAutomovelUseCase.ts
@@ -1,3 +1,4 @@
+import { IAutomovelRepository } from "@modules/automovel/repositories/IAutomovelRepository";
 import { ITipoAutomovelRepository } from "@modules/automovel/repositories/ITipoAutomovelRepository";
 import { AppError } from "@shared/errors/AppError";
 import { inject, injectable } from "tsyringe";
@@ -6,13 +7,19 @@ import { inject, injectable } from "tsyringe";
 class DeleteTipoAutomovelUseCase {
     constructor(
         @inject("TipoAutomovelRepository")
-        private tipoAutomovelRepository: ITipoAutomovelRepository
+        private tipoAutomovelRepository: ITipoAutomovelRepository,
+
+        @inject("AutomovelRepository")
+        private automovelRepository: IAutomovelRepository
     ) {}
     
     async execute(id: string): Promise<void> {
-        this.tipoAutomovelRepository.delete(id);
-    }
+        const tipoAutomovelIsUsed = await this.automovelRepository.findByType(id)
 
+        if (tipoAutomovelIsUsed.length > 0) throw new AppError('Registro não pode ser excluído pois existe(m) Automóvel(is) cadastrado(s) com este Tipo Automóvel!')
+        
+        this.tipoAutomovelRepository.delete(id)
+    }
 }
 
 export { DeleteTipoAutomovelUseCase };
